Backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
python-jose[cryptography]==3.3.0
cryptography==43.0.3
pyotp==2.9.0
python-multipart==0.0.9